Вынес подключени к MongoDB в отдельный файл
