+++ appconfig.js
@@ -1,4 +1,3 @@
-const mongoose = require('mongoose');
 const rateLimit = require('express-rate-limit');
 
 const PORT = process.env.PORT || 3000;
@@ -6,21 +5,6 @@ const { NODE_ENV } = process.env;
 const isProduction = process.env.NODE_ENV === 'production'; // if production, we use secret from env file
 const JWT_SECRET = isProduction ? process.env.JWT_SECRET : 'devSecretKey'; // for dev mode, we use 'devSecretKey'
 
-
-// подключение MongoDB
-const connectToMongoDB = (async () => {
-  try {
-    await mongoose.connect('mongodb://localhost:27017/explorer-api', {
-      useNewUrlParser: true,
-      useCreateIndex: true,
-      useFindAndModify: false,
-      useUnifiedTopology: true,
-    });
-  } catch (err) {
-    console.error(`Failed to connect to MongoDB: ${err.message}, ${err.stack}`);
-    process.exit(1);
-  }
-});
 
 // настройки cookie
 const JWT_COOKIE_OPTIONS = {
@@ -39,7 +23,6 @@ const rateLimiter = rateLimit({
 module.exports = {
   PORT,
   JWT_SECRET,
-  connectToMongoDB,
   rateLimiter,
   JWT_COOKIE_OPTIONS,
 };
